fix range errors

+++ src/mutf8.nim
@@ -4,13 +4,13 @@ proc decodeMUTF8*(s: string): string =
   let stringStream: StringStream = newStringStream(s)
   let length: int = len(s)
 
-  var bytes: seq[int8] = @[]
+  var bytes: seq[int] = @[]
 
   while not stringStream.atEnd():
     bytes.add(stringStream.readInt8())
 
   var i: int = 0
-  var x: int8
+  var x: int
 
   while i < length:
     x = bytes[i]
@@ -31,11 +31,12 @@ proc decodeMUTF8*(s: string): string =
       if i + 5 > length:
         raise newException(ValueError, "6-byte codepoint started, but too short to finish.")
 
-      x = int8(0x10000 or
-        int((bytes[i+1] and 0b00001111) shl 0x10) or
-        int((bytes[i+2] and 0b00111111) shl 0x0A) or
-        int((bytes[i+4] and 0b00001111) shl 0x06) or
-        int(bytes[i+5] and 0b00111111)
+      x = (
+        0x10000 or
+        ((bytes[i+1] and 0b00001111) shl 0x10) or
+        ((bytes[i+2] and 0b00111111) shl 0x0A) or
+        ((bytes[i+4] and 0b00001111) shl 0x06) or
+        (bytes[i+5] and 0b00111111)
       )
 
       i += 6
